requirement.txt
asgiref==3.5.2
Django==4.1.4
django-leaflet==0.28.3
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
